test_tour_Tour: Move setup into separate util_ files
The test is already getting very unweildy, and likely to get so
unweildy that it needs breaking into multiple files soon.

Break out the test setup into it's own util to improve both.
